Swift: Use Bulk middleware to do delete_multi when it is available
Before this commit *delete_multi* would raise a *NoSuchObject* error when *force* was *False* and some keys were not found. Now it swallows this error in any case. This change was necessary since we cannot find out which keys were not found from the Bulk middleware response.

The doc string of *AbstractBackend.delete_multi* already documents this possibility.

--HG--
branch : swift-feature-detection1