Opportunistically remove blocks earlier
When removing multiple blocks, first attempt to remove blocks for
which the lock can be immediately acquired. This is important when
removing a file right after it has been created, because if the upload
of the first block starts before removal the removal thread will
otherwise be stuck behind the upload procedure, waiting for every
block to be uploaded only to remove it afterwards.