Fix size record of dirty blocks uploaded by fsck.
Since commit 06ac78a4 (2020-03-22) we have accidentally rounded up the size of dirty
blocks to the next multiple of 512 when uploading them during fsck. This means that
any inode with such a block as the last one was deemed to have inplausible size,
resulting in fsck.s3ql then extending the size and thereby creating a hole with
zero-bytes at the end of the file.