Merge 'blocks' and 'objects' tables
Remove the distinction between blocks and objects. Until now, S3QL had the theoretical
capability to store multiple blocks in the same backend object. However, this was never
fully supported by code and therefore association between blocks and objects was always
1:1.

It does not look like this is going to change in the future, so by removing the
distinction we reduce the size of the metadata table, improve performance, and make the
code easier to understand.

Still todo:

 - Update file system code
 - Write migration code

Fixes: #62